backend/pyproject.toml
[project]
name = "grantfiller-api"
version = "0.1.0"
description = "Local grant drafting API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "aiosqlite>=0.20.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.12",
    "pymupdf>=1.24.0",
    "python-docx>=1.1.2",
    "fpdf2>=2.8.0",
    "rank-bm25>=0.2.2",
    "trafilatura>=1.8.0",
    "certifi>=2024.0.0",
    "playwright>=1.49.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
